feat(modding): type node-rec layout fields as WIT enums
display/position-type/overflow/flex-direction/justify-content/align-items cross
as WIT enums instead of raw u8 — mods get typed variants instead of magic
numbers. Case order matches the host Val.cs enums so the converter casts the
discriminant straight across. Bumps the wasmtime-dotnet submodule for the
enum-in-record unsafe-context source-gen fix.

+++ src/TinyEcs.Bevy.Modding/TypedComponents.cs
@@ -21,12 +21,15 @@ internal static class ModTypedComponents
 
     private static Node ToNode(in WitApp.NodeRec n) => new()
     {
-        Display = (Display)n.Display,
-        PositionType = (PositionType)n.PositionType,
-        Overflow = (Overflow)n.Overflow,
-        FlexDirection = (FlexDirection)n.FlexDirection,
-        JustifyContent = (JustifyContent)n.JustifyContent,
-        AlignItems = (AlignItems)n.AlignItems,
+        // WIT enum discriminant -> host enum. WIT case order == host enum order
+        // (both 0-based, Val.cs), so cast through the integral value. C# forbids
+        // a direct enum->enum cast.
+        Display = (Display)(byte)n.Display,
+        PositionType = (PositionType)(byte)n.PositionType,
+        Overflow = (Overflow)(byte)n.Overflow,
+        FlexDirection = (FlexDirection)(byte)n.FlexDirection,
+        JustifyContent = (JustifyContent)(byte)n.JustifyContent,
+        AlignItems = (AlignItems)(byte)n.AlignItems,
         Width = ToVal(n.Width), Height = ToVal(n.Height),
         MinWidth = ToVal(n.MinWidth), MinHeight = ToVal(n.MinHeight),
         MaxWidth = ToVal(n.MaxWidth), MaxHeight = ToVal(n.MaxHeight),
